# PERF: Optimize Implementation2.cs