@joseftw 2022-03-14
Cool! However, I don't think anyone should use inplementation 1-2, only inplementation 3. Do you agree? ;)